+++ database/migrations/2026_04_24_000006_webmail_provider_roundcube.php
@@ -11,6 +11,7 @@ return function (PDO $pdo): void {
         'mail_webmail_imap_host' => '',
         'mail_webmail_smtp_host' => '',
         'mail_webmail_doc_root' => '',
+        'mail_webmail_aliases' => '[]',
         'mail_webmail_install_task_id' => '',
         'mail_webmail_install_status' => '',
         'mail_webmail_installed_at' => '',